+++ CHANGELOG.md
@@ -7,6 +7,11 @@
 - **Breaking change:** Tasks can no longer take a `GrinderContext` parameter.
   Instead, they must access the `context` getter.
 
+- **Breaking change:** Remove deprecated members `startGrinder()`, `task()`,
+  `runProcess()`, `runProcessAsync()`, `getSdkDir()`, and
+  `Analyzer.analyzeFiles()`. Remove the deprecated `quiet` argument to
+  `Dart.version()`.
+
 - **Breaking change:** Remove `package:grinder/grinder_files.dart`,
   `package:grinder/grinder_sdk.dart`, and `package:grinder/grinder_tools.dart`.
   All members provided by these libraries are still available through the main